# apps/docs/openapi/routess.openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/routes": {
			"post": {
				"description": "Creates a new route for the authenticated user. PATs are blocked; route creation lands with #170.",
				"operationId": "RoutesController_create",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRouteDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Route created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid route data"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Create a new route",
				"tags": ["routes"]
			},
			"get": {
				"description": "Retrieves the authenticated user's routes (any visibility), newest first. Paginated via `limit` and `offset`; the total number of routes is returned in the `X-Total-Count` response header.",
				"operationId": "RoutesController_findAll",
				"parameters": [
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"description": "Page size",
						"schema": {
							"minimum": 1,
							"maximum": 200,
							"default": 100,
							"type": "number"
						}
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"description": "Number of routes to skip",
						"schema": {
							"minimum": 0,
							"default": 0,
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Routes retrieved successfully. X-Total-Count carries the total route count.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/RouteResponseDto"
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [
					{
						"PAT-auth": []
					},
					{
						"JWT-auth": []
					}
				],
				"summary": "Get all user routes",
				"tags": ["routes"]
			}
		},
...
		"/routes/{id}": {
			"get": {
				"description": "Returns the route. Owners see it regardless of visibility; non-owners (including anonymous viewers) only see public and unlisted routes. Private routes return 404 to non-owners. PATs hit this through the cookie-or-bearer JWT path; PAT-as-Bearer is not yet supported here (use GET /routes to list and filter).",
				"operationId": "RoutesController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Route ID",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Route retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Route not found"
					}
				},
				"summary": "Get route by ID",
				"tags": ["routes"]
			},
			"patch": {
				"description": "Updates a specific route for the authenticated user.",
				"operationId": "RoutesController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Route ID",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "X-Routess-Confirm",
						"in": "header",
						"description": "Set to `true` when a PAT call sets `visibility: public`. Cookie sessions ignore this header.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRouteDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Route updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid route data"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Route not found"
					},
					"428": {
						"description": "PAT-authenticated call attempted to set visibility to public without `X-Routess-Confirm: true`. Surface the `impact` field of the response to the user and retry with the header set."
					}
				},
				"security": [
					{
						"PAT-auth": []
					},
					{
						"JWT-auth": []
					}
				],
				"summary": "Update route",
				"tags": ["routes"]
			},
			"delete": {
				"description": "Deletes a specific route for the authenticated user. PAT callers must set X-Routess-Confirm: true.",
				"operationId": "RoutesController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Route ID",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "X-Routess-Confirm",
						"in": "header",
						"description": "Required as `true` for PAT callers. Cookie sessions ignore this header.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Route deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"example": {
										"success": true,
										"message": "Route deleted successfully"
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Route not found"
					},
					"428": {
						"description": "PAT-authenticated DELETE without `X-Routess-Confirm: true`. Surface the `impact` field of the response to the user and retry with the header set."
					}
				},
				"security": [
					{
						"PAT-auth": []
					},
					{
						"JWT-auth": []
					}
				],
				"summary": "Delete route",
				"tags": ["routes"]
			}
		},
...
		"/admin/stats/overview": {
			"get": {
				"description": "Returns headline counts (users, routes, signups) computed from Postgres aggregates. Backs the admin overview page.",
				"operationId": "AdminController_getOverview",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Aggregate top-level metrics for the admin dashboard",
				"tags": ["admin"]
			}
		},
		"/admin/stats/users": {
			"get": {
				"description": "Signup trend, churn, and active-user counts computed from Postgres aggregates.",
				"operationId": "AdminController_getUserStats",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "User-level KPIs (signups, churn, active count)",
				"tags": ["admin"]
			}
		},
		"/admin/stats/routes": {
			"get": {
				"description": "Route KPIs grouped by activity, the most active creators, and the recent creation trend.",
				"operationId": "AdminController_getRouteStats",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Route counts by activity, top creators, recent trend",
				"tags": ["admin"]
			}
		},
		"/admin/routes": {
			"get": {
				"description": "Lists routes across all users with `page`/`pageSize` pagination, free-text `search`, and an optional `userId` filter.",
				"operationId": "AdminController_listRoutes",
				"parameters": [
					{
						"name": "search",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "userId",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Paginated, searchable route list",
				"tags": ["admin"]
			}
		},
		"/admin/routes/{id}": {
			"get": {
				"description": "Full admin view of one route, including its owner and a summary of its waypoints.",
				"operationId": "AdminController_getRouteDetail",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Route detail: metadata, waypoints summary, owner",
				"tags": ["admin"]
			},
			"delete": {
				"description": "Marks the route deleted without erasing the row. The owner no longer sees it.",
				"operationId": "AdminController_softDeleteRoute",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"204": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Soft-delete a route",
				"tags": ["admin"]
			}
		},
		"/admin/users": {
			"get": {
				"description": "Lists users with `page`/`pageSize` pagination and free-text `search` over name and email.",
				"operationId": "AdminController_listUsers",
				"parameters": [
					{
						"name": "search",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Paginated, searchable user list",
				"tags": ["admin"]
			}
		},
		"/admin/users/{id}": {
			"get": {
				"description": "Full admin view of one user: profile fields, active sessions, and their most recent routes.",
				"operationId": "AdminController_getUserDetail",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "User detail: profile, sessions, recent routes",
				"tags": ["admin"]
			},
			"delete": {
				"description": "Marks the user deleted, soft-deletes their routes, and revokes all their sessions.",
				"operationId": "AdminController_softDeleteUser",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"204": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Soft-delete a user; cascades to routes and sessions",
				"tags": ["admin"]
			}
		},
		"/admin/users/{id}/sessions/{sessionId}": {
			"delete": {
				"description": "Force-signs the user out of one device by revoking that session.",
				"operationId": "AdminController_revokeSession",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "sessionId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Revoke a single session for a user",
				"tags": ["admin"]
			}
		},
...
		"/admin/system/config-summary": {
			"get": {
				"description": "Shows which optional integrations are configured (email, Valhalla, Umami, Sentry) without exposing secret values.",
				"operationId": "AdminController_getConfigSummary",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Sanitised view of runtime config (no secrets)",
				"tags": ["admin"]
			}
		}
	},
	"info": {
		"title": "Routess API",
		"description": "API for route management and mapping functionality",
		"version": "dev",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "https://api.routess.com",
			"description": "Production"
		},
		{
			"url": "http://localhost:3000",
			"description": "Local development"
		}
	],
	"components": {
		"securitySchemes": {
			"JWT-auth": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http",
				"in": "header"
			},
			"PAT-auth": {
				"scheme": "bearer",
				"bearerFormat": "routess_pat_<random>",
				"type": "http",
				"description": "Personal access token for non-browser clients. Mint at Settings → API Tokens. Carries one of two scopes: `read` (list/get Routes, export GPX, profile) or `write` (`read` plus metadata-only mutations on owned routes and on preferences). Blocked from /admin/* and from DELETE /users/me. Destructive operations require `X-Routess-Confirm: true` and return 428 PRECONDITION_REQUIRED otherwise. See ADR-0022 and ADR-0023.",
				"in": "header"
			}
		},
		"schemas": {
...
			"RoutingPreferencesDto": {
				"type": "object",
				"properties": {
					"surfacePreference": {
						"type": "string",
						"enum": ["paved", "mixed", "unpaved"],
						"example": "mixed"
					},
					"avoidFerries": {
						"type": "boolean",
						"example": true
					},
					"avoidHighways": {
						"type": "boolean",
						"example": true
					}
				},
				"required": ["surfacePreference", "avoidFerries", "avoidHighways"]
			},
...
			"WaypointDto": {
				"type": "object",
				"properties": {
					"coord": {
						"type": "array",
						"description": "Waypoint coordinate as a [lng, lat] pair",
						"example": [-74.006, 40.7128],
						"items": {
							"type": "number"
						},
						"minItems": 2,
						"maxItems": 2
					},
					"type": {
						"type": "string",
						"description": "Type of waypoint routing",
						"enum": ["routed", "direct"],
						"example": "routed"
					},
					"name": {
						"type": "string",
						"description": "Optional user-assigned name for the waypoint",
						"example": "Coffee stop"
					},
					"timestamp": {
						"type": "string",
						"description": "Timestamp when the waypoint was recorded",
						"example": "2024-01-15T10:30:00Z"
					}
				},
				"required": ["coord", "type"]
			},
			"CreateRouteDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"description": "Name of the route",
						"example": "Morning Jog Route",
						"minLength": 1
					},
					"description": {
						"type": "string",
						"description": "Optional description of the route",
						"example": "A scenic route through Central Park"
					},
					"activity": {
						"type": "string",
						"description": "Activity associated with the route",
						"enum": ["run", "cycle", "walk"]
					},
					"visibility": {
						"type": "string",
						"description": "Visibility of the route",
						"enum": ["private", "unlisted", "public"],
						"default": "private"
					},
					"tags": {
						"description": "Free-form tags for the route",
						"example": ["hilly", "scenic"],
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"favourite": {
						"type": "boolean",
						"description": "Whether the owner marked this route as a favourite",
						"default": false
					},
					"waypoints": {
						"description": "Array of waypoints that define the route",
						"minItems": 1,
						"example": [
							{
								"coord": [-74.006, 40.7128],
								"type": "routed",
								"timestamp": "2024-01-15T10:00:00Z"
							},
							{
								"coord": [-73.9851, 40.7589],
								"type": "direct",
								"timestamp": "2024-01-15T10:30:00Z"
							}
						],
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/WaypointDto"
						}
					},
					"geometry": {
						"description": "Routed polyline as an array of [lng, lat] coordinate pairs",
						"example": [[-74.006, 40.7128], [-74.0055, 40.7135]],
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"distance": {
						"type": "number",
						"description": "Total distance of the route in meters",
						"example": 5280
					},
					"duration": {
						"type": "number",
						"description": "Estimated duration of the route in seconds",
						"example": 1200
					},
					"elevationGain": {
						"type": "number",
						"description": "Total elevation gain of the route in meters",
						"example": 150
					},
					"startAddress": {
						"type": "string",
						"description": "Human-readable starting address",
						"example": "Central Park, New York, NY"
					},
					"endAddress": {
						"type": "string",
						"description": "Human-readable ending address",
						"example": "Times Square, New York, NY"
					},
					"routingPreferences": {
						"description": "Inputs that produced this route's geometry. Optional: routes saved from clients without this metadata (e.g. legacy clients, GPX imports, agents that didn't capture the prefs) are persisted with `routingPreferences = null` and remain usable but cannot be recalculated against their original inputs.",
						"allOf": [
							{
								"$ref": "#/components/schemas/RoutingPreferencesDto"
							}
						]
					},
					"provenance": {
						"type": "string",
						"description": "How this route was produced. Defaults to 'valhalla' for new routes.",
						"enum": ["valhalla", "mapbox-legacy", "gpx-import", "generation"]
					}
				},
				"required": ["name", "waypoints"]
			},
...
			"RouteResponseDto": {
				"type": "object",
				"properties": {
					"legs": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/RouteLegDto"
						}
					},
					"locations": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/RouteSnappedLocationDto"
						}
					}
				},
				"required": ["legs", "locations"]
			},
			"UpdateRouteDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"description": "Name of the route",
						"example": "Morning Jog Route",
						"minLength": 1
					},
					"description": {
						"type": "string",
						"description": "Optional description of the route",
						"example": "A scenic route through Central Park"
					},
					"activity": {
						"type": "string",
						"description": "Activity associated with the route",
						"enum": ["run", "cycle", "walk"]
					},
					"visibility": {
						"type": "string",
						"description": "Visibility of the route",
						"enum": ["private", "unlisted", "public"],
						"default": "private"
					},
					"tags": {
						"description": "Free-form tags for the route",
						"example": ["hilly", "scenic"],
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"favourite": {
						"type": "boolean",
						"description": "Whether the owner marked this route as a favourite",
						"default": false
					},
					"waypoints": {
						"description": "Array of waypoints that define the route",
						"minItems": 1,
						"example": [
							{
								"coord": [-74.006, 40.7128],
								"type": "routed",
								"timestamp": "2024-01-15T10:00:00Z"
							},
							{
								"coord": [-73.9851, 40.7589],
								"type": "direct",
								"timestamp": "2024-01-15T10:30:00Z"
							}
						],
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/WaypointDto"
						}
					},
					"geometry": {
						"description": "Routed polyline as an array of [lng, lat] coordinate pairs",
						"example": [[-74.006, 40.7128], [-74.0055, 40.7135]],
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"distance": {
						"type": "number",
						"description": "Total distance of the route in meters",
						"example": 5280
					},
					"duration": {
						"type": "number",
						"description": "Estimated duration of the route in seconds",
						"example": 1200
					},
					"elevationGain": {
						"type": "number",
						"description": "Total elevation gain of the route in meters",
						"example": 150
					},
					"startAddress": {
						"type": "string",
						"description": "Human-readable starting address",
						"example": "Central Park, New York, NY"
					},
					"endAddress": {
						"type": "string",
						"description": "Human-readable ending address",
						"example": "Times Square, New York, NY"
					},
					"routingPreferences": {
						"description": "Inputs that produced this route's geometry. Optional: routes saved from clients without this metadata (e.g. legacy clients, GPX imports, agents that didn't capture the prefs) are persisted with `routingPreferences = null` and remain usable but cannot be recalculated against their original inputs.",
						"allOf": [
							{
								"$ref": "#/components/schemas/RoutingPreferencesDto"
							}
						]
					}
				}
			},
...
			"RouteLegSummaryDto": {
				"type": "object",
				"properties": {
					"length": {
						"type": "number",
						"description": "Leg length in kilometers."
					},
					"time": {
						"type": "number",
						"description": "Leg duration in seconds."
					}
				},
				"required": ["length", "time"]
			},
			"RouteLegDto": {
				"type": "object",
				"properties": {
					"shape": {
						"type": "string",
						"description": "Polyline6-encoded shape of the leg."
					},
					"summary": {
						"$ref": "#/components/schemas/RouteLegSummaryDto"
					}
				},
				"required": ["shape", "summary"]
			},
			"RouteSnappedLocationDto": {
				"type": "object",
				"properties": {
					"lat": {
						"type": "number",
						"description": "Snapped latitude (may differ from input)."
					},
					"lon": {
						"type": "number",
						"description": "Snapped longitude (may differ from input)."
					},
					"original_index": {
						"type": "number",
						"description": "Index of the originally-supplied location, when Valhalla returns it."
					}
				},
				"required": ["lat", "lon"]
			}
		}
	}
}